% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.R
\name{add_keywords}
\alias{add_keywords}
\title{Add keywords to a DataStore reference}
\usage{
add_keywords(
  reference_id,
  keywords,
  dev = TRUE,
  interactive = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{reference_id}{Numeric reference ID. You must have the appropriate permissions to edit this reference.}

\item{keywords}{A character vector of keywords}

\item{dev}{Logical. Defaults to TRUE because it's best to attempt to modify references on the development & testing version of DataStore first. When everything is working, change to \code{dev = FALSE} and run again to edit the real reference.}

\item{interactive}{Logical. Prompt for user confirmation before uploading?}

\item{verbose}{Logical. Leave this option as FALSE unless you have reported a bug and are asked to provide diagnostic info.}
}
\value{
A character vector of all keywords for the reference
}
\description{
Add keywords to a DataStore reference
}
\examples{
\dontrun{
my_keywords <- c("bison", "human-wildlife conflict", "visitor injuries")
all_keywords <- add_keywords(reference_id = 00000, keywords = my_keywords, dev = TRUE)
}

}
