% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.R
\name{add_producing_units}
\alias{add_producing_units}
\title{Add producing units to a DataStore reference}
\usage{
add_producing_units(reference_id, nps_units, dev = TRUE, interactive = TRUE)
}
\arguments{
\item{reference_id}{Numeric reference ID. You must have the appropriate permissions to edit this reference.}

\item{nps_units}{A character vector of park/unit codes}

\item{dev}{Logical. Defaults to TRUE because it's best to attempt to modify references on the development & testing version of DataStore first. When everything is working, change to \code{dev = FALSE} and run again to edit the real reference.}

\item{interactive}{Logical. Prompt for user confirmation before uploading?}
}
\value{
A character vector of all unit codes for the reference
}
\description{
Add producing units to a DataStore reference
}
\examples{
\dontrun{
my_producing_units <- c("LAKE", "DEVA", "MOJA", "JOTR")
all_producing_units <- add_producing_units(reference_id = 00000, parks = my_content_units, dev = TRUE)
}

}
