% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.R
\name{create_draft_reference}
\alias{create_draft_reference}
\title{Create a new draft reference}
\usage{
create_draft_reference(
  title,
  reference_type_code,
  date_published,
  date_precision_code,
  dev = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{title}{The title of the reference}

\item{reference_type_code}{The code indicating the type of reference. See \code{get_reference_types()} for a full list of valid reference types.}

\item{date_published}{Text string with the date of publication. MUST be in "YYYY-MM-DD", "YYYY-MM", or "YYYY" format. Often the current date, but may be earlier if the resource was previously published outside of DataStore (often the case with journal articles, newsletters, etc.).}

\item{date_precision_code}{Optional. If \code{date_published} only specifies the year, you may optionally use this to specify a season or quarter. See \code{get_date_precision()} for valid options.}

\item{dev}{Logical. Defaults to TRUE because it's best to attempt to modify references on the development & testing version of DataStore first. When everything is working, change to \code{dev = FALSE} and run again to edit the real reference.}

\item{verbose}{Logical. Leave this option as FALSE unless you have reported a bug and are asked to provide diagnostic info.}
}
\value{
A list containing information about the uploaded reference, including reference profile URL and reference ID.
}
\description{
Create a new draft reference
}
\examples{
\dontrun{
  new_ref <- create_draft_reference(title = "2024 Oakridge Goose Inventory",
                                    reference_type_code = "Datapackage",
                                    date_published = "2025-02-14",
                                    dev = TRUE)

  another_new_ref <- create_draft_reference(title = "Oakridge Wildlife Hazards Update: Spring 2025",
                                            reference_type_code = "NewsletterArticle",
                                            date_published = "2025",
                                            date_precision_code = "Spring",
                                            dev = TRUE)
}


}
