% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{datastore_example_refs}
\alias{datastore_example_refs}
\title{Retrieve some valid DataStore reference IDs}
\usage{
datastore_example_refs(visiblity = c("public", "internal", "both"), n, seed)
}
\arguments{
\item{visiblity}{Return public-facing IDs, internal-facing IDs, or a mix of both?}

\item{n}{Optional. Number of IDs to return. Will randomly sample \code{n} IDs. If \code{visibility == 'both'}, there is no guarantee that the sample of \code{n} IDs will return both public and internal-facing IDs. There are 48 public-facing IDs and 45 internal-facing.}

\item{seed}{Optional. Set this if you need your sample of IDs to come back the same every time.}
}
\value{
An integer vector of DataStore reference IDs.
}
\description{
For example and testing purposes. See \verb{?NPSdatastore::public_refs and ?NPSdatastore::internal_refs} for more information.
}
\examples{
public_ids <- datastore_example_refs(visibility = "public", n = 5)
all_ids <- datastore_example_refs(visibility = "both")

}
