% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifecycle.R
\name{delete_inactive_ref}
\alias{delete_inactive_ref}
\title{Delete an inactive reference}
\usage{
delete_inactive_ref(reference_id, dev = TRUE, interactive = TRUE)
}
\arguments{
\item{reference_id}{Numeric reference ID. You must have the appropriate permissions to edit this reference.}

\item{dev}{Logical. Defaults to TRUE because it's best to attempt to modify references on the development & testing version of DataStore first. When everything is working, change to \code{dev = FALSE} and run again to edit the real reference.}

\item{interactive}{Logical. Prompt for user confirmation before uploading?}
}
\value{
NULL (invisibly)
}
\description{
This function will delete one reference. The reference must be in an inactive status and the person running the function must be an editor on the reference to delete it. Use with caution. This is not easily undone!
}
\examples{
\dontrun{
delete_inactive_ref(reference_id = 00000,
                    dev = TRUE,
                    interactive = TRUE)
}
}
