% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete.R
\name{delete_reference_owner}
\alias{delete_reference_owner}
\title{Delete an owner from a DataStore reference}
\usage{
delete_reference_owner(
  reference_id,
  upn,
  email,
  dev = TRUE,
  interactive = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{reference_id}{Numeric reference ID. You must have the appropriate permissions to edit this reference.}

\item{upn}{A single user's UPN. Must specify either \code{upn} or \code{email} but not both.}

\item{email}{A single user's email. Must specify either \code{upn} or \code{email} but not both.}

\item{dev}{Logical. Defaults to TRUE because it's best to attempt to modify references on the development & testing version of DataStore first. When everything is working, change to \code{dev = FALSE} and run again to edit the real reference.}

\item{interactive}{Logical. Prompt for user confirmation before uploading?}

\item{verbose}{Logical. Leave this option as FALSE unless you have reported a bug and are asked to provide diagnostic info.}
}
\value{
A tibble of all reference owners with columns userCode, lastName, firstName, and email
}
\description{
Delete an owner from a DataStore reference
}
\examples{
\dontrun{

all_owners <- delete_reference_owner(reference_id = 00000,
                                     email = "edward_abbey@nps.gov",
                                     dev = TRUE)
all_owners <- delete_reference_owner(reference_id = 00000,
                                     upn = "gmwright@nps.gov",
                                     dev = TRUE)
}

}
