% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.datastore_request}
\alias{.datastore_request}
\title{Create httr2 request for DataStore API}
\usage{
.datastore_request(is_secure, is_dev, suppress_errors = TRUE, verbose = FALSE)
}
\arguments{
\item{is_secure}{Retrieve the secure version of the API base URL?}

\item{is_dev}{Retrieve the dev version of the API base URL?}

\item{suppress_errors}{Suppress HTTP errors? Set to TRUE if using \code{.validate_resp()}}

\item{verbose}{Logical. Leave this option as FALSE unless you have reported a bug and are asked to provide diagnostic info.}
}
\value{
A httr2 request object with curl options set to allow authentication for NPS users (if using secure API)
}
\description{
Create httr2 request for DataStore API
}
\keyword{internal}
