% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.get_base_units_url}
\alias{.get_base_units_url}
\title{Get the right base URL for the Units API}
\usage{
.get_base_units_url(is_dev)
}
\arguments{
\item{is_dev}{Retrieve the dev version of the units base API?}
}
\value{
One of two base URL for Units API (production, dev)
}
\description{
Get the right base URL for the Units API
}
\keyword{internal}
