% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.get_base_url}
\alias{.get_base_url}
\title{Get the right base URL for the DataStore API}
\usage{
.get_base_url(is_secure, is_dev)
}
\arguments{
\item{is_secure}{Retrieve the secure version of the API base URL?}

\item{is_dev}{Retrieve the dev version of the API base URL?}
}
\value{
One of four base URLs for the DataStore API (public, secure, public+dev, secure+dev)
}
\description{
Get the right base URL for the DataStore API
}
\keyword{internal}
