% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.get_ref_profile_url}
\alias{.get_ref_profile_url}
\title{Given a reference ID, construct the URL to its profile page}
\usage{
.get_ref_profile_url(ref_id, is_dev)
}
\arguments{
\item{ref_id}{A reference ID}

\item{is_dev}{Retrieve the dev version of the API base URL?}
}
\value{
The url to the reference profile page
}
\description{
Given a reference ID, construct the URL to its profile page
}
