% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.get_reference_profiles}
\alias{.get_reference_profiles}
\title{Perform a single request to the Profile endpoint and tidy the data a little}
\usage{
.get_reference_profiles(reference_ids, is_secure, is_dev, verbose = FALSE)
}
\arguments{
\item{reference_ids}{numeric vector of <=25 reference IDs}

\item{is_secure}{Retrieve the secure version of the API base URL?}

\item{is_dev}{Retrieve the dev version of the API base URL?}
}
\value{
List of reference profiles
}
\description{
This endpoint only returns 25 results at a time; this helper function is used
inside of a loop or apply fxn to support retrieval of >25 profiles at a time
}
\keyword{internal}
