% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.lists2tibbles}
\alias{.lists2tibbles}
\title{Convert lists in a reference profile to tibbles}
\usage{
.lists2tibbles(parent_list, child_list_names)
}
\arguments{
\item{parent_list}{The list representing the reference profile}

\item{child_list_names}{The elements of the reference profile that should be converted to tibbles}
}
\value{
The tidied reference profile
}
\description{
Use for reference profile elements like taxa and units that are returned as
lists but would be more appropriately stored as tibbles
}
