% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.validate_ref_id}
\alias{.validate_ref_id}
\title{Validate reference ID arguments}
\usage{
.validate_ref_id(
  ref_id,
  multiple_ok = FALSE,
  arg = rlang::caller_arg(ref_id),
  call = rlang::caller_env()
)
}
\arguments{
\item{ref_id}{The reference ID argument to validate}

\item{multiple_ok}{Can ref_id be a vector of multiple IDs?}

\item{arg}{Used to get the actual name of the argument in the calling function. See \verb{?rlang::}topic-error-call``}

\item{call}{The caller environment, for more helpful error messages. See \verb{?rlang::}topic-error-call``}
}
\description{
Validate reference ID arguments
}
\keyword{internal}
