% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.validate_resp}
\alias{.validate_resp}
\title{Check HTTP response and throw a helpful error if needed}
\usage{
.validate_resp(
  resp,
  nice_msg_400,
  nice_msg_500,
  details,
  call = rlang::caller_env()
)
}
\arguments{
\item{resp}{HTTP response as returned by \code{httr2::req_perform()}}

\item{nice_msg_400}{Optional. Custom message for errors in the 400 range}

\item{nice_msg_500}{Optional. Custom message for errors in the 500 range}

\item{details}{Optional. Name of JSON field in response body that contains more info about the error}

\item{call}{Caller environment}
}
\description{
Check HTTP response and throw a helpful error if needed
}
\keyword{internal}
