% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.validate_truefalse}
\alias{.validate_truefalse}
\title{Validate TRUE/FALSE arguments}
\usage{
.validate_truefalse(
  bool,
  arg = rlang::caller_arg(bool),
  call = rlang::caller_env()
)
}
\arguments{
\item{bool}{Value to check}

\item{arg}{Used to get the actual name of the argument in the calling function. See \verb{?rlang::}topic-error-call``}

\item{call}{The caller environment, for more helpful error messages. See \verb{?rlang::}topic-error-call``}
}
\description{
Validate TRUE/FALSE arguments
}
