% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units_service.R
\name{get_all_nps_units}
\alias{get_all_nps_units}
\title{Provides a list of all National Park Service Units}
\usage{
get_all_nps_units(dev = FALSE, verbose = FALSE)
}
\arguments{
\item{dev}{Logical. Set to TRUE to use the development & testing version of the API. If this means nothing to you, ignore and allow it to default to FALSE.}

\item{verbose}{Logical. Leave this option as FALSE unless you have reported a bug and are asked to provide diagnostic info.}
}
\value{
a tibble where each row is an NPS unit
}
\description{
Provides a list of all National Park Service Units
}
\examples{
\dontrun{
units <- get_all_nps_units()
}
}
