% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_lists.R
\name{get_legal_authority}
\alias{get_legal_authority}
\title{Get a list of legal authorities for restricting DataStore reference files}
\usage{
get_legal_authority(dev = FALSE, verbose = FALSE)
}
\arguments{
\item{dev}{Logical. Set to TRUE to use the development & testing version of the API. If this means nothing to you, ignore and allow it to default to FALSE.}

\item{verbose}{Logical. Leave this option as FALSE unless you have reported a bug and are asked to provide diagnostic info.}
}
\value{
A tibble where each row has a separate legal authority for restricting file downloads on DataStore.
}
\description{
The function \code{get_legal_authority} returns a tibble of the current acceptable legal authorities for restricting files attached to references on DataStore.
}
\examples{
 \dontrun{
authorities <- get_legal_authority()
}
}
