% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_lifecycle_info}
\alias{get_lifecycle_info}
\title{Retrieve the current lifecycle status of a reference.
Only works for internal NPS users.}
\usage{
get_lifecycle_info(reference_id, dev = FALSE, verbose = FALSE)
}
\arguments{
\item{reference_id}{Numeric reference ID. You must have the appropriate permissions to edit this reference.}

\item{dev}{Logical. Defaults to TRUE because it's best to attempt to modify references on the development & testing version of DataStore first. When everything is working, change to \code{dev = FALSE} and run again to edit the real reference.}

\item{verbose}{Logical. Leave this option as FALSE unless you have reported a bug and are asked to provide diagnostic info.}
}
\value{
A list of lifecycle information for the reference
}
\description{
Retrieve the current lifecycle status of a reference.
Only works for internal NPS users.
}
\examples{
\dontrun{
  lifecycle_info <- get_lifecycle_info(reference_id = 652358)
  lifecycle = lifecycle_info$lifecycle
}

}
