% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_reference_owners}
\alias{get_reference_owners}
\title{List the owners of a DataStore reference}
\usage{
get_reference_owners(reference_id, dev = FALSE)
}
\arguments{
\item{reference_id}{Numeric reference ID. You must have the appropriate permissions to edit this reference.}

\item{dev}{Logical. Defaults to TRUE because it's best to attempt to modify references on the development & testing version of DataStore first. When everything is working, change to \code{dev = FALSE} and run again to edit the real reference.}
}
\value{
A tibble with the username, last name, first name, and email of each reference owner.
}
\description{
Only available for NPS users on the internal network.
}
\examples{
\dontrun{
owners <- get_reference_owners(reference_id = 652358)
}

}
