% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units_service.R
\name{get_unit_geography}
\alias{get_unit_geography}
\title{Get unit geography}
\usage{
get_unit_geography(
  units,
  detail = c("convexHull", "envelope", "feature"),
  dataformat = c("wkt", "glm"),
  dev = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{units}{String, List. One or more valid units (e.g. "ROMO" or c("ROMO", "YElL"))}

\item{detail}{String. Defaults to "convexHull". Acceptable alternatives are "envelope" or "feature".}

\item{dataformat}{String. Defaults to "wkt". An acceptable alternative is "glm"}

\item{dev}{Logical. Defaults to FALSE. Should the dev API be used?}

\item{verbose}{Logical. Defaults to FALSE. Should lots of information about the API call be printed to the console? Great for debugging.}
}
\value{
A tibble of units and their geography
}
\description{
Get unit geography
}
\examples{
\dontrun{
get_unit_geography("ROMO")
get_unit_geography(c("ROMO", "YELL"))
}
}
