% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{internal_refs}
\alias{internal_refs}
\title{Reference IDs for internal-facing references}
\format{
\subsection{\code{internal_refs}}{

A vector of 45 reference IDs
}
}
\source{
DataStore advanced search for "spinus pinus" in taxonomy. Search performed 03/26/2025.
}
\usage{
internal_refs
}
\description{
A small subset of reference IDs corresponding to internal-facing DataStore reference
profiles. To be used for examples and testing.
}
\keyword{datasets}
