% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{public_refs}
\alias{public_refs}
\title{Reference IDs for public-facing references}
\format{
\subsection{\code{public_refs}}{

A vector of 48 reference IDs
}
}
\source{
DataStore advanced search for "spinus pinus" in taxonomy. Search performed 03/26/2025.
}
\usage{
public_refs
}
\description{
A small subset of reference IDs corresponding to public-facing DataStore reference
profiles. To be used for examples and testing.
}
\keyword{datasets}
