% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/high_level.R
\name{set_by_for_nps_list}
\alias{set_by_for_nps_list}
\title{Sets the "by or for NPS" flag for multiple references}
\usage{
set_by_for_nps_list(
  reference_id,
  by_for_nps = TRUE,
  life_cycle = c("Status Quo", "Active", "Draft"),
  nps_internal = TRUE,
  dev = TRUE
)
}
\arguments{
\item{reference_id}{Integer. Typically a seven-digit code}

\item{by_for_nps}{Logical. Defaults to TRUE}

\item{life_cycle}{String. Defaults to "Status Quo". Acceptable alternatives are "Active" and "Draft".}

\item{nps_internal}{Logical. Defaults to TRUE. Must be TRUE for most write actions.}

\item{dev}{Logical. Target the development server? Defaults to TRUE. False will target the production server.}
}
\value{
a dataframe
}
\description{
This function allows the user to change the "by or for NPS" flag for one or more references. It also gives the user the option of making all the references active, putting them all in a draft status, or leaving the status of each reference the way it originally was.
}
\details{
This function includes relatively little error checking and gives no opportunity for user input once it has started. The idea is that the user has done all the vetting prior to using the function. For instance, if a reference cannot be activated for some reason, the function will end in an error.

Common errors include: supplying the wrong reference id number, not being connected to the VPN or on an NPS network, not having the correct permissions to make changes to the reference (e.g. you are not an owner of the reference) and, if attempting to activate a reference not having supplied all of the information necessary to activate the reference (perhaps there are no files attached or perhaps you have not specified any authors).

The function returns a dataframe with the reference id, the current (after running the function) status of the "by or for NPS", the initial and final lifecycle status of each reference, and a hyperlink to the reference profile.
}
\examples{
\dontrun{
 nps_status <- set_by_for_nps_list(reference_id = c(1234567,
                                                    7654321,
                                                    0000000),
                                   by_for_nps = TRUE,
                                   life_cycle = "Status Quo",
                                   nps_internal = TRUE,
                                   dev = TRUE)
                                   }
}
