% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.R
\name{set_license}
\alias{set_license}
\title{Set the license for a DataStore reference}
\usage{
set_license(
  reference_id,
  license_type_id,
  dev = TRUE,
  interactive = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{reference_id}{Numeric reference ID. You must have the appropriate permissions to edit this reference.}

\item{license_type_id}{The numeric ID of the license you wish to apply.}

\item{dev}{Logical. Defaults to TRUE because it's best to attempt to modify references on the development & testing version of DataStore first. When everything is working, change to \code{dev = FALSE} and run again to edit the real reference.}

\item{interactive}{Logical. Prompt for user confirmation before uploading?}

\item{verbose}{Logical. Leave this option as FALSE unless you have reported a bug and are asked to provide diagnostic info.}
}
\value{
A list of the referenceId and the licenseTypeId
}
\description{
Any existing license will be overwritten.
}
\examples{
\dontrun{
new_bib <- add_license(000000,
                       license_id = 1)  # CC0 license
}

}
