% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify.R
\name{set_reference_owners}
\alias{set_reference_owners}
\title{Add owners to a DataStore reference}
\usage{
set_reference_owners(
  reference_id,
  upns,
  emails,
  dev = TRUE,
  interactive = TRUE
)
}
\arguments{
\item{reference_id}{Numeric reference ID. You must have the appropriate permissions to edit this reference.}

\item{upns}{A character vector of UPNs}

\item{emails}{A character vector of email addresses}

\item{dev}{Logical. Defaults to TRUE because it's best to attempt to modify references on the development & testing version of DataStore first. When everything is working, change to \code{dev = FALSE} and run again to edit the real reference.}

\item{interactive}{Logical. Prompt for user confirmation before uploading?}
}
\value{
A tibble of all reference owners with columns userCode, lastName, firstName, and email
}
\description{
Add owners to a DataStore reference
}
\examples{
\dontrun{
upns <- c("gmwright@nps.gov", "ymexia@nps.gov")
emails <- c("enid_michael@nps.gov", "edward_abbey@nps.gov")

all_owners <- set_reference_owners(reference_id = 00000, upns = upns, emails = emails, dev = TRUE)

}

}
