% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{active_directory_lookup}
\alias{active_directory_lookup}
\title{Look up emails and/or UPNs in Active Directory}
\usage{
active_directory_lookup(upns, emails, verbose = FALSE)
}
\arguments{
\item{upns}{A character vector of UPNs}

\item{emails}{A character vector of email addresses}
}
\value{
A dataframe of user information with a row for each UPN and email address searched, regardless of whether it exists in the system. If \code{found} is \code{TRUE}, the user exists. If \code{disabled} is \code{FALSE}, the user exists but has been deactivated.
}
\description{
Only works for NPS users on the internal network. Accepts a vector of UPNs, a vector of emails, or both.
}
\examples{
\dontrun{
upns <- c("gmwright@nps.gov", "ymexia@nps.gov")
emails <- c("enid_michael@nps.gov", "edward_abbey@nps.gov")

all_users <- active_directory_lookup(upns, emails)
emails_only <- active_directory_lookup(emails = emails)
}

}
