% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.R
\name{add_content_units}
\alias{add_content_units}
\title{Add content units to a DataStore reference}
\usage{
add_content_units(
  reference_id,
  parks,
  include_linked_units = FALSE,
  add_bounding_box = TRUE,
  dev = TRUE,
  interactive = TRUE
)
}
\arguments{
\item{reference_id}{Numeric reference ID. You must have the appropriate permissions to edit this reference.}

\item{parks}{A character vector of park/unit codes}

\item{dev}{Logical. Defaults to TRUE because it's best to attempt to modify references on the development & testing version of DataStore first. When everything is working, change to \code{dev = FALSE} and run again to edit the real reference.}

\item{interactive}{Logical. Prompt for user confirmation before uploading?}
}
\value{
A character vector of all unit codes for the reference
}
\description{
Add content units to a DataStore reference
}
\examples{
\dontrun{
my_content_units <- c("LAKE", "DEVA", "MOJA", "JOTR")
all_content_units <- add_content_units(reference_id = 00000, parks = my_content_units, dev = TRUE)
}

}
