% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.R
\name{add_to_project}
\alias{add_to_project}
\title{Add DataStore reference(s) to a Project reference}
\usage{
add_to_project(project_id, reference_ids, dev = TRUE, interactive = TRUE)
}
\arguments{
\item{project_id}{The reference ID of the Project}

\item{reference_ids}{Numeric vector of reference IDs.}

\item{dev}{Logical. Set to TRUE to use the development & testing version of the API. If this means nothing to you, ignore and allow it to default to FALSE.}

\item{interactive}{Logical. Prompt for user confirmation before uploading?}
}
\value{
A character vector of all keywords for the reference
}
\description{
Add DataStore reference(s) to a Project reference
}
\examples{
\dontrun{
  proj_id <- 1111111
  ref_ids <- c(000000, 222222)
  add_to_project(project_id = proj_id,
                 reference_ids = ref_ids,
                 dev = TRUE)
}

}
