% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete.R
\name{delete_all_keywords}
\alias{delete_all_keywords}
\title{Delete all keywords from a DataStore reference}
\usage{
delete_all_keywords(reference_id, dev = TRUE, interactive = TRUE)
}
\arguments{
\item{reference_id}{Numeric reference ID. You must have the appropriate permissions to edit this reference.}

\item{dev}{Logical. Defaults to TRUE because it's best to attempt to modify references on the development & testing version of DataStore first. When everything is working, change to \code{dev = FALSE} and run again to edit the real reference.}

\item{interactive}{Logical. Prompt for user confirmation before uploading?}
}
\value{
Invisibly returns the reference ID
}
\description{
Delete all keywords from a DataStore reference
}
\examples{
\dontrun{
delete_all_keywords(reference_id = 00000, dev = TRUE)
}

}
