% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_by_for_nps}
\alias{get_by_for_nps}
\title{Determine if a reference was created by/for the NPS}
\usage{
get_by_for_nps(reference_id, nps_internal = FALSE, dev = FALSE)
}
\arguments{
\item{reference_id}{Numeric reference ID. You must have the appropriate permissions to edit this reference.}

\item{nps_internal}{Logical. NPS users on the internal network can set to TRUE to authenticate and view non-public data and create or modify references. All other users can ignore this argument and allow it to default to FALSE.}

\item{dev}{Logical. Defaults to TRUE because it's best to attempt to modify references on the development & testing version of DataStore first. When everything is working, change to \code{dev = FALSE} and run again to edit the real reference.}
}
\value{
TRUE if the reference was created by or for the NPS, FALSE if not.
}
\description{
Determine if a reference was created by/for the NPS
}
\examples{
\dontrun{
  by_for_nps <- get_by_for_nps(reference_id = 652358)
}

}
