% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup.R
\name{get_date_precision}
\alias{get_date_precision}
\title{Get a list of valid values for date precision}
\usage{
get_date_precision(dev = FALSE)
}
\arguments{
\item{dev}{Logical. Set to TRUE to use the development & testing version of the API. If this means nothing to you, ignore and allow it to default to FALSE.}
}
\value{
A tibble with columns for date precision code and label
}
\description{
Get a list of valid values for date precision
}
\examples{
valid_precisions <- get_date_precision()

}
