% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_file_info}
\alias{get_file_info}
\title{Retrieve information about files associated with a DataStore reference}
\usage{
get_file_info(reference_id, file_id, nps_internal = FALSE, dev = FALSE)
}
\arguments{
\item{reference_id}{Numeric reference ID. You must have the appropriate permissions to edit this reference.}

\item{file_id}{Optional. Omit to retrieve info for all files in the reference. Unique ID (sometimes referred to as "resource ID" in DataStore) corresponding to the file you would like information about.}

\item{nps_internal}{Logical. NPS users on the internal network can set to TRUE to authenticate and view non-public data and create or modify references. All other users can ignore this argument and allow it to default to FALSE.}

\item{dev}{Logical. Defaults to TRUE because it's best to attempt to modify references on the development & testing version of DataStore first. When everything is working, change to \code{dev = FALSE} and run again to edit the real reference.}
}
\value{
A tibble with columns userSort, resourceId, lastUpdate, description, fileName, fileSize_kb, extension, mimeType, downloadLink, and is508Compliant. Returns a zero-length tibble if no files exist.
}
\description{
Retrieve information about files associated with a DataStore reference
}
\examples{
all_files <- get_file_info(reference_id = 652358)
specific_file <- get_file_info(reference_id = 2305163, file_id = 706913)

}
