% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup.R
\name{get_keywords}
\alias{get_keywords}
\title{Retrieve the keywords for a DataStore reference}
\usage{
get_keywords(reference_id, nps_internal = FALSE, dev = FALSE)
}
\arguments{
\item{reference_id}{Numeric. The reference ID for a single DataStore reference.}

\item{nps_internal}{Logical. NPS users on the internal network can set to TRUE to authenticate and view non-public data and create or modify references. All other users can ignore this argument and allow it to default to FALSE.}

\item{dev}{Logical. Set to TRUE to use the development & testing version of the API. If this means nothing to you, ignore and allow it to default to FALSE.}
}
\value{
A character vector of keywords.
}
\description{
Retrieve the keywords for a DataStore reference
}
\examples{
owners <- get_keywords(reference_id = 652358)

}
