% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_keywords}
\alias{get_keywords}
\title{Retrieve the keywords for a DataStore reference}
\usage{
get_keywords(reference_id, nps_internal = FALSE, dev = FALSE)
}
\arguments{
\item{reference_id}{Numeric reference ID. You must have the appropriate permissions to edit this reference.}

\item{nps_internal}{Logical. NPS users on the internal network can set to TRUE to authenticate and view non-public data and create or modify references. All other users can ignore this argument and allow it to default to FALSE.}

\item{dev}{Logical. Defaults to TRUE because it's best to attempt to modify references on the development & testing version of DataStore first. When everything is working, change to \code{dev = FALSE} and run again to edit the real reference.}
}
\value{
A character vector of keywords.
}
\description{
Retrieve the keywords for a DataStore reference
}
\examples{
owners <- get_keywords(reference_id = 652358)

}
