% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup.R
\name{get_reference_owners}
\alias{get_reference_owners}
\title{List the owners of a DataStore reference}
\usage{
get_reference_owners(reference_id, dev = FALSE)
}
\arguments{
\item{reference_id}{Numeric. The reference ID for a single DataStore reference.}

\item{dev}{Logical. Set to TRUE to use the development & testing version of the API. If this means nothing to you, ignore and allow it to default to FALSE.}
}
\value{
A tibble with the username, last name, first name, and email of each reference owner.
}
\description{
Only available for NPS users on the internal network.
}
\examples{
owners <- get_reference_owners(reference_id = 652358)

}
