% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{search_references_by_id}
\alias{search_references_by_id}
\title{Search NPS DataStore references by their ID}
\usage{
search_references_by_id(reference_ids, nps_internal = FALSE, dev = FALSE)
}
\arguments{
\item{reference_ids}{Numeric vector of reference IDs. The API only allows up to 25 reference IDs at a time.}

\item{nps_internal}{Logical. NPS users on the internal network can set to TRUE to authenticate and view non-public data and create or modify references. All other users can ignore this argument and allow it to default to FALSE.}

\item{dev}{Logical. Set to TRUE to use the development & testing version of the API. If this means nothing to you, ignore and allow it to default to FALSE.}
}
\value{
fill in later
}
\description{
Search NPS DataStore references by their ID
}
