% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{search_references_by_id}
\alias{search_references_by_id}
\title{Search NPS DataStore references by their ID}
\usage{
search_references_by_id(reference_ids, nps_internal = FALSE, dev = FALSE)
}
\arguments{
\item{reference_ids}{Numeric vector of reference IDs.}

\item{nps_internal}{Logical. NPS users on the internal network can set to TRUE to authenticate and view non-public data and create or modify references. All other users can ignore this argument and allow it to default to FALSE.}

\item{dev}{Logical. Set to TRUE to use the development & testing version of the API. If this means nothing to you, ignore and allow it to default to FALSE.}
}
\value{
A named list containing detailed information about each reference, where the names are the reference IDs.
}
\description{
Search NPS DataStore references by their ID
}
\examples{
\dontrun{
ids <- example_ref_ids(n = 2, seed = 2025)
ref_profiles <- search_references_by_id(reference_ids = ids,
                                        nps_internal = FALSE,
                                        dev = FALSE)
}

}
