% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{search_references_by_id_basic}
\alias{search_references_by_id_basic}
\title{Search NPS DataStore references by ID}
\usage{
search_references_by_id_basic(reference_ids, nps_internal = FALSE, dev = FALSE)
}
\arguments{
\item{reference_ids}{Numeric vector of reference IDs.}

\item{nps_internal}{Logical. NPS users on the internal network can set to TRUE to authenticate and view non-public data and create or modify references. All other users can ignore this argument and allow it to default to FALSE.}

\item{dev}{Logical. Set to TRUE to use the development & testing version of the API. If this means nothing to you, ignore and allow it to default to FALSE.}
}
\value{
A tibble containing basic information about each reference.
}
\description{
Retrieve only basic information about each reference. To return more detailed information about references, use \code{search_references_by_id()} instead.
}
\examples{
\dontrun{
ids <- example_ref_ids(n = 2, seed = 2025)
ref_profiles <- search_references_by_id_basic(reference_ids = ids,
                                        nps_internal = FALSE,
                                        dev = FALSE)
}

}
