% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.R
\name{set_bibliography}
\alias{set_bibliography}
\title{Replace the bibliography in a DataStore reference}
\usage{
set_bibliography(reference_id, bibliography, dev = TRUE, interactive = TRUE)
}
\arguments{
\item{reference_id}{Numeric reference ID. You must have the appropriate permissions to edit this reference.}

\item{bibliography}{A list representing a DataStore bibliography. It is recommended that you retrieve the current bibliography using \code{get_bibliography()} and then modify it as needed.}

\item{dev}{Logical. Defaults to TRUE because it's best to attempt to modify references on the development & testing version of DataStore first. When everything is working, change to \code{dev = FALSE} and run again to edit the real reference.}

\item{interactive}{Logical. Prompt for user confirmation before uploading?}
}
\value{
A list representing the new bibliography.
}
\description{
To update specific elements of the bibliography, use the "set_" functions for those elements instead.
}
\examples{
\dontrun{
bib <- get_bibliography(reference_id = 00000)
bib$description <- "This is a new description for this reference"
bib$notes <- "This reference is for testing purposes only"
new_bib <- set_bibliography(reference_id = 00000, bibliography = bib, dev = TRUE)
}

}
