% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.R
\name{set_by_for_nps}
\alias{set_by_for_nps}
\title{Set the flag that indicates whether a reference was created by or for NPS}
\usage{
set_by_for_nps(reference_id, by_for_nps, dev = TRUE, interactive = TRUE)
}
\arguments{
\item{reference_id}{Numeric reference ID. You must have the appropriate permissions to edit this reference.}

\item{by_for_nps}{TRUE or FALSE: was this reference created by or for NPS?}

\item{dev}{Logical. Defaults to TRUE because it's best to attempt to modify references on the development & testing version of DataStore first. When everything is working, change to \code{dev = FALSE} and run again to edit the real reference.}

\item{interactive}{Logical. Prompt for user confirmation before uploading?}
}
\value{
A list representing the full updated bibliography.
}
\description{
Set the flag that indicates whether a reference was created by or for NPS
}
\examples{
\dontrun{
new_bib <- set_by_for_nps(reference_id = 000000, by_for_nps = TRUE)
new_bib <- set_by_for_nps(reference_id = 000000, by_for_nps = TRUE, dev = FALSE)
}
}
