% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.R
\name{set_file_info}
\alias{set_file_info}
\title{Set file information}
\usage{
set_file_info(
  reference_id,
  file_id,
  description,
  is_508,
  dev = TRUE,
  interactive = TRUE
)
}
\arguments{
\item{reference_id}{Numeric reference ID. You must have the appropriate permissions to edit this reference.}

\item{file_id}{The unique identifier for the file. Also referred to as the resource ID.}

\item{description}{A short description of the file}

\item{is_508}{Is the file 508 compliant?}

\item{dev}{Logical. Defaults to TRUE because it's best to attempt to modify references on the development & testing version of DataStore first. When everything is working, change to \code{dev = FALSE} and run again to edit the real reference.}

\item{interactive}{Logical. Prompt for user confirmation before uploading?}
}
\value{
A tibble containing the updated file information.
}
\description{
Set file information
}
\examples{
\dontrun{
  set_file_info(2313985, file_id = 731902, description = "testing 123")
}

}
